requirements.txt
flask==3.1.0
openai==1.68.0
python-dotenv==1.1.0
playwright==1.50.0
PyJWT==2.9.0
Flask-Limiter==3.8.0
fastapi==0.115.0
uvicorn[standard]==0.30.0
pydantic==2.9.0
httpx==0.27.0
google-api-python-client==2.130.0
APScheduler==3.10.4
